move to ArrayPathAccessor
